Refactor UI into scan and history modules
