+++ ui/__init__.py
@@ -1,1 +1,12 @@
-# UI package
+from .scan import render_scanner_tab
+from .history import render_history_tab
+from .debugger import render_debugger_tab
+from .layout import setup_page, render_header
+
+__all__ = [
+    "render_scanner_tab",
+    "render_history_tab",
+    "render_debugger_tab",
+    "setup_page",
+    "render_header",
+]
